Bug 1445091 Part 1 - Support temporary snapshots in record/replay interface, no_r.

+++ toolkit/recordreplay/ipc/Channel.cpp
@@ -222,14 +222,14 @@ PrintMessage(const char* aPrefix, const Message& aMsg)
   case MessageType::HitSnapshot: {
     const HitSnapshotMessage& nmsg = *(const HitSnapshotMessage*)&aMsg;
     data = new char[128];
-    snprintf(data, 128, "Id %zu, Final %d, Interim %d",
-             nmsg.mSnapshotId, nmsg.mFinal, nmsg.mInterim);
+    snprintf(data, 128, "Id %d, Final %d, Interim %d",
+             (int) nmsg.mSnapshotId, nmsg.mFinal, nmsg.mInterim);
     break;
   }
   case MessageType::HitBreakpoint: {
     const HitBreakpointMessage& nmsg = *(const HitBreakpointMessage*)&aMsg;
     data = new char[128];
-    snprintf(data, 128, "Id %zu", nmsg.mBreakpointId);
+    snprintf(data, 128, "Id %d", (int) nmsg.mBreakpointId);
     break;
   }
   case MessageType::Resume: {
